timeutils: move format functions to a separate module
The prototypes of these functions are not yet changed, except for naming,
which is not ideal, as we should conform to a common pattern there.

But the patch is large enough already, so I just renamed the functions
for now.
