+++ lib/logwriter.c
@@ -36,6 +36,7 @@
 #include "ml-batched-timer.h"
 #include "str-format.h"
 #include "scratch-buffers.h"
+#include "timeutils/format.h"
 
 #include <unistd.h>
 #include <assert.h>
@@ -936,7 +937,7 @@ log_writer_format_log(LogWriter *self, LogMessage *lm, GString *result)
       g_string_append_c(result, '1');
       g_string_append_c(result, ' ');
 
-      unix_time_append_format(stamp, result, TS_FMT_ISO,
+      append_format_unix_time(stamp, result, TS_FMT_ISO,
                               time_zone_info_get_offset(self->options->template_options.time_zone_info[LTZ_SEND], stamp->ut_sec),
                               self->options->template_options.frac_digits);
       g_string_append_c(result, ' ');
@@ -1014,7 +1015,7 @@ log_writer_format_log(LogWriter *self, LogMessage *lm, GString *result)
 
           if (self->flags & LW_FORMAT_FILE)
             {
-              unix_time_format(stamp, result, self->options->template_options.ts_format,
+              format_unix_time(stamp, result, self->options->template_options.ts_format,
                                time_zone_info_get_offset(self->options->template_options.time_zone_info[LTZ_SEND], stamp->ut_sec),
                                self->options->template_options.frac_digits);
             }
@@ -1025,7 +1026,7 @@ log_writer_format_log(LogWriter *self, LogMessage *lm, GString *result)
               g_string_append_c(result, '>');
 
               /* always use BSD timestamp by default, the use can override this using a custom template */
-              unix_time_append_format(stamp, result, TS_FMT_BSD,
+              append_format_unix_time(stamp, result, TS_FMT_BSD,
                                       time_zone_info_get_offset(self->options->template_options.time_zone_info[LTZ_SEND], stamp->ut_sec),
                                       self->options->template_options.frac_digits);
             }
