+++ lib/timeutils/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(TIMEUTILS_HEADERS
     timeutils/cache.h
     timeutils/scan-timestamp.h
+    timeutils/format.h
     timeutils/names.h
     timeutils/strptime-tz.h
     timeutils/timeutils.h
@@ -12,6 +13,7 @@ set(TIMEUTILS_HEADERS
 set(TIMEUTILS_SOURCES
     timeutils/cache.c
     timeutils/scan-timestamp.c
+    timeutils/format.c
     timeutils/names.c
     timeutils/strptime-tz.c
     timeutils/timeutils.c
