+++ modules/afuser/afuser.c
@@ -24,6 +24,7 @@
 #include "afuser.h"
 #include "messages.h"
 #include "compat/getutent.h"
+#include "timeutils/format.h"
 
 #include <string.h>
 #include <fcntl.h>
@@ -108,7 +109,7 @@ afuser_dd_queue(LogPipe *s, LogMessage *msg, const LogPathOptions *path_options)
     goto finish;
 
   timestamp = g_string_sized_new(0);
-  unix_time_format(&msg->timestamps[LM_TS_STAMP], timestamp, TS_FMT_FULL, -1, 0);
+  format_unix_time(&msg->timestamps[LM_TS_STAMP], timestamp, TS_FMT_FULL, -1, 0);
   g_snprintf(buf, sizeof(buf), "%s %s %s\n",
              timestamp->str,
              log_msg_get_value(msg, LM_V_HOST, NULL),
