+++ tests/unit/test_zone.c
@@ -26,6 +26,7 @@
 #include "timeutils/cache.h"
 #include "timeutils/zoneinfo.h"
 #include "timeutils/unixtime.h"
+#include "timeutils/format.h"
 #include <criterion/criterion.h>
 #include <stdlib.h>
 #include <string.h>
@@ -122,7 +123,7 @@ assert_time_zone_offset(TimezoneOffsetTestCase c)
 void
 assert_timestamp_format(GString *target, UnixTime *stamp, TimestampFormatTestCase c)
 {
-  unix_time_format(stamp, target, c.format, c.zone_offset, c.frac_digits);
+  format_unix_time(stamp, target, c.format, c.zone_offset, c.frac_digits);
   cr_assert_str_eq(target->str, c.expected_format, "Actual: %s, Expected: %s", target->str, c.expected_format);
 }
